Remove auto filtering option
This XR setting is only useful on native Quest apps,
to allow dynamically changing the sharpening/supersampling filter
depending on the performance of the game being played.
This does not affect PCVR at all.

Also added warnings relative to the fact that MQSR superseded SGSR,
while still allowing users to enable it manually if they so choose.